Commands are now clean and fixed
