+++ src/main/java/net/tardis/mod/Tardis.java
@@ -1,8 +1,5 @@
 package net.tardis.mod;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
 import net.minecraft.init.Items;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -34,108 +31,27 @@ import net.tardis.mod.client.models.exteriors.TileEntityDoorTT;
 import net.tardis.mod.common.ars.ConsoleRoom;
 import net.tardis.mod.common.blocks.TBlocks;
 import net.tardis.mod.common.commands.CommandDebug;
-import net.tardis.mod.common.commands.CommandTardis;
+import net.tardis.mod.common.commands.TardisCommand;
 import net.tardis.mod.common.dimensions.TDimensions;
 import net.tardis.mod.common.dimensions.TDimensions.BiomeReg;
-import net.tardis.mod.common.entities.EntityAdipose;
-import net.tardis.mod.common.entities.EntityChair;
-import net.tardis.mod.common.entities.EntityCompanion;
-import net.tardis.mod.common.entities.EntityCorridor;
-import net.tardis.mod.common.entities.EntityCybermanInvasion;
-import net.tardis.mod.common.entities.EntityDalek;
-import net.tardis.mod.common.entities.EntityDalekCasing;
-import net.tardis.mod.common.entities.EntityDalekSkaro;
-import net.tardis.mod.common.entities.EntityDummy;
-import net.tardis.mod.common.entities.EntityItemMaterializer;
-import net.tardis.mod.common.entities.EntityLaserRay;
-import net.tardis.mod.common.entities.EntityQuark;
-import net.tardis.mod.common.entities.EntityRaider;
+import net.tardis.mod.common.entities.*;
 import net.tardis.mod.common.entities.brak.EntityDoorsBrakSecondary;
-import net.tardis.mod.common.entities.controls.ControlDimChange;
-import net.tardis.mod.common.entities.controls.ControlDirection;
-import net.tardis.mod.common.entities.controls.ControlDoor;
-import net.tardis.mod.common.entities.controls.ControlDoorSwitch;
-import net.tardis.mod.common.entities.controls.ControlFastReturn;
-import net.tardis.mod.common.entities.controls.ControlFuel;
-import net.tardis.mod.common.entities.controls.ControlLandType;
-import net.tardis.mod.common.entities.controls.ControlLaunch;
-import net.tardis.mod.common.entities.controls.ControlMag;
-import net.tardis.mod.common.entities.controls.ControlMonitor;
-import net.tardis.mod.common.entities.controls.ControlPhone;
-import net.tardis.mod.common.entities.controls.ControlRandom;
-import net.tardis.mod.common.entities.controls.ControlSonicSlot;
-import net.tardis.mod.common.entities.controls.ControlStabilizers;
-import net.tardis.mod.common.entities.controls.ControlTelepathicCircuts;
-import net.tardis.mod.common.entities.controls.ControlWaypoint;
-import net.tardis.mod.common.entities.controls.ControlX;
-import net.tardis.mod.common.entities.controls.ControlY;
-import net.tardis.mod.common.entities.controls.ControlZ;
+import net.tardis.mod.common.entities.controls.*;
 import net.tardis.mod.common.entities.hellbent.EntityHellbentCorridor;
 import net.tardis.mod.common.entities.hellbent.EntityHellbentDoor;
 import net.tardis.mod.common.entities.vehicles.EntityBessie;
 import net.tardis.mod.common.items.ItemSonic;
 import net.tardis.mod.common.items.TItems;
-import net.tardis.mod.common.protocols.ProtocolCCircuit;
-import net.tardis.mod.common.protocols.ProtocolChangeInterior;
-import net.tardis.mod.common.protocols.ProtocolConsole;
-import net.tardis.mod.common.protocols.ProtocolEnabledHADS;
-import net.tardis.mod.common.protocols.ProtocolFindDimDRfit;
-import net.tardis.mod.common.protocols.ProtocolRepair;
-import net.tardis.mod.common.protocols.ProtocolStealth;
-import net.tardis.mod.common.protocols.ProtocolSystemReadout;
-import net.tardis.mod.common.protocols.ProtocolToggleHum;
-import net.tardis.mod.common.protocols.ProtocolWaypoints;
-import net.tardis.mod.common.protocols.TardisProtocol;
+import net.tardis.mod.common.protocols.*;
 import net.tardis.mod.common.recipes.RepairRecipes;
 import net.tardis.mod.common.screwdriver.ScrewdriverHandler;
 import net.tardis.mod.common.strings.TStrings;
-import net.tardis.mod.common.systems.SystemAntenna;
-import net.tardis.mod.common.systems.SystemCCircuit;
-import net.tardis.mod.common.systems.SystemDimension;
-import net.tardis.mod.common.systems.SystemFlight;
-import net.tardis.mod.common.systems.SystemFluidLinks;
-import net.tardis.mod.common.systems.SystemStabilizers;
-import net.tardis.mod.common.systems.SystemTemporalGrace;
-import net.tardis.mod.common.systems.SystemThermo;
-import net.tardis.mod.common.systems.TardisSystems;
-import net.tardis.mod.common.tileentity.TileEntityAlembic;
+import net.tardis.mod.common.systems.*;
+import net.tardis.mod.common.tileentity.*;
 import net.tardis.mod.common.tileentity.TileEntityAlembic.AlembicRecipe;
-import net.tardis.mod.common.tileentity.TileEntityComponentRepair;
-import net.tardis.mod.common.tileentity.TileEntityDoor;
-import net.tardis.mod.common.tileentity.TileEntityEPanel;
-import net.tardis.mod.common.tileentity.TileEntityEgg;
-import net.tardis.mod.common.tileentity.TileEntityFoodMachine;
-import net.tardis.mod.common.tileentity.TileEntityHellbentLight;
-import net.tardis.mod.common.tileentity.TileEntityItemMaterializer;
-import net.tardis.mod.common.tileentity.TileEntityJsonTester;
-import net.tardis.mod.common.tileentity.TileEntityKerblam;
-import net.tardis.mod.common.tileentity.TileEntityLight;
-import net.tardis.mod.common.tileentity.TileEntityMultiblock;
-import net.tardis.mod.common.tileentity.TileEntityMultiblockMaster;
-import net.tardis.mod.common.tileentity.TileEntitySonicGun;
-import net.tardis.mod.common.tileentity.TileEntitySonicWorkbench;
-import net.tardis.mod.common.tileentity.TileEntityTardis;
-import net.tardis.mod.common.tileentity.TileEntityTardisCoral;
-import net.tardis.mod.common.tileentity.TileEntityUmbrellaStand;
-import net.tardis.mod.common.tileentity.consoles.TileEntityTardis01;
-import net.tardis.mod.common.tileentity.consoles.TileEntityTardis02;
-import net.tardis.mod.common.tileentity.consoles.TileEntityTardis03;
-import net.tardis.mod.common.tileentity.consoles.TileEntityTardis04;
-import net.tardis.mod.common.tileentity.consoles.TileEntityTardis05;
-import net.tardis.mod.common.tileentity.decoration.TileEntityAmSphere;
-import net.tardis.mod.common.tileentity.decoration.TileEntityChair;
-import net.tardis.mod.common.tileentity.decoration.TileEntityHelbentRoof;
-import net.tardis.mod.common.tileentity.decoration.TileEntityHellbentMonitor;
-import net.tardis.mod.common.tileentity.decoration.TileEntityHellbentPole;
-import net.tardis.mod.common.tileentity.decoration.TileEntityToyotaSpin;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoor01;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoor03;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoor04;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoor05;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoorCC;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoorClock;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoorWardrobe;
-import net.tardis.mod.common.tileentity.exteriors.TileEntityDoorWood;
+import net.tardis.mod.common.tileentity.consoles.*;
+import net.tardis.mod.common.tileentity.decoration.*;
+import net.tardis.mod.common.tileentity.exteriors.*;
 import net.tardis.mod.common.world.TardisLoadingCallback;
 import net.tardis.mod.common.world.WorldGenTardis;
 import net.tardis.mod.config.TardisConfig;
@@ -145,6 +61,8 @@ import net.tardis.mod.network.NetworkHandler;
 import net.tardis.mod.proxy.ServerProxy;
 import net.tardis.mod.util.common.helpers.EntityHelper;
 import net.tardis.mod.util.common.helpers.FileHelper;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 
 @Mod(modid = Tardis.MODID, name = Tardis.NAME, version = Tardis.VERSION, dependencies = Tardis.DEP, updateJSON = Tardis.UPDATE_JSON_URL)
@@ -354,7 +272,6 @@ public class Tardis {
 		PermissionAPI.registerNode(TStrings.Permissions.RESTORE_TARDIS, DefaultPermissionLevel.OP, "Allows players to restore their TARDIS Systems");
 		PermissionAPI.registerNode(TStrings.Permissions.GROW, DefaultPermissionLevel.OP, "Allows players to grow their TARDIS Coral faster");
 		PermissionAPI.registerNode(TStrings.Permissions.TP_OUT_TARDIS, DefaultPermissionLevel.OP, "Allows players to teleport themself out of their TARDIS");
-		PermissionAPI.registerNode(TStrings.Permissions.TP_OUT_TARDIS_OTHER, DefaultPermissionLevel.OP, "Allows players to teleport themself out of TARDIS of a specified player");
 
 		//This should be in pre-init, but it seems some mods have a weird obsession with claiming already taken ids
 		TDimensions.register();
@@ -373,7 +290,7 @@ public class Tardis {
 
 	@EventHandler
 	public void serverStarting(FMLServerStartingEvent event) {
-		event.registerServerCommand(new CommandTardis());
+		event.registerServerCommand(new TardisCommand());
 		if (Tardis.getIsDev()) {
 			event.registerServerCommand(new CommandDebug());
 		}
