+++ src/main/java/net/tardis/mod/common/strings/TStrings.java
@@ -74,7 +74,6 @@ public class TStrings {
 		public static final String TP_IN_TARDIS = "tardis.command.tardistp";
 		public static final String TP_IN_TARDIS_OTHER = "tardis.command.tardistp.other";
 		public static final String TP_OUT_TARDIS = "tardis.command.exterior";
-		public static final String TP_OUT_TARDIS_OTHER = "tardis.command.exterior.other";
 		public static final String REMOVE_TARDIS = "tardis.command.removetardis";
 		public static final String RESTORE_TARDIS = "tardis.command.restoretardis";
 		public static final String GROW = "tardis.command.grow";
